### 2019-12-18 #1967 @whedon
[PRE REVIEW]: genius: analysis ready song lyric retrieval
**Submitting author:** @josiahparry (<a href="http://orcid.org/0000-0001-9910-865X">Josiah Parry</a>)
**Repository:** <a href="https://github.com/josiahparry/genius" target ="_blank">https://github.com/josiahparry/genius</a>
**Version:** 2.2.1
