### 2019-12-18 #1967 @JosiahParry
I'll get to adding the missing DOIs.

Potential Reviewers:
@alexhanna: R user + NLP
@Faroit: is a pyton user but seems to do work in MIR
