+++ paper/index.bib
@@ -35,6 +35,7 @@ url="https://doi.org/10.1007/978-3-642-11674-2_15"
   journal={Cognition and Emotion},
   pages={1--14},
   year={2019},
+  doi="10.1080/02699931.2019.1700482",
   publisher={Taylor \& Francis}
 }
 
@@ -84,6 +85,7 @@ title = {Songs perceived as relaxing: Musical features, lyrics, and contributing
   number={4},
   pages={161--176},
   year={2018},
+  doi="10.1525/jpms.2018.300411",
   publisher={University of California Press Journals}
 }
 
@@ -94,7 +96,7 @@ title = {Songs perceived as relaxing: Musical features, lyrics, and contributing
   note = {R package version 2.1.1},
   url = {https://CRAN.R-project.org/package=spotifyr},
   }
-  
+
 @article{tidytext,
   title={tidytext: Text Mining and Analysis Using Tidy Data Principles in R.},
   author={Silge, Julia and Robinson, David},
@@ -102,7 +104,8 @@ title = {Songs perceived as relaxing: Musical features, lyrics, and contributing
   volume={1},
   number={3},
   pages={37},
-  year={2016}
+  year={2016},
+  doi="10.21105/joss.00037"
 }
 
 @article{tidydata,
@@ -113,6 +116,6 @@ title = {Songs perceived as relaxing: Musical features, lyrics, and contributing
   number={10},
   pages={1--23},
   year={2014},
+  doi="10.18637/jss.v059.i10",
   publisher={Foundation for Open Access Statistics}
 }
-  
