### 2019-12-29 #1967 @faroit
@danielskatz @JosiahParry I could do it but I am no expert on R software and also I work mainly with audio whereas this contribution is more focussed on text metadata if I'm not mistaken.
