### 2019-12-29 #1967 @faroit
Maybe consider @bmcfee he's also a python head but has significantly more experience with core MIR
