### 2019-12-30 #1967 @bmcfee
Sure @danielskatz , thanks for the invitation!  I'm not too much of an R user, so it might not be a bad idea to get an R person to chime in as well.
