### 2020-01-06 #1967 @faroit
One more suggestions: I know @heraldoborges who is an active R user and a researcher in Machine Learning/Information Retrieval.
