### 2020-01-06 #1967 @danielskatz
👋 @heraldoborges - would you be willing to review this for JOSS?  (If you haven't reviewed for JOSS before, please see our [intro to reviewing](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html))
