### 2020-01-06 #1967 @heraldoborges
Yes.

Em seg, 6 de jan de 2020 3:25 PM, Daniel S. Katz <notifications@github.com>
escreveu:

> 👋 @heraldoborges <https://github.com/heraldoborges> - would you be
> willing to review this for JOSS? (If you haven't reviewed for JOSS before,
> please see our intro to reviewing
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html> and review
> criteria <https://joss.readthedocs.io/en/latest/review_criteria.html>)
>
