### 2020-01-06 #1995 @danielskatz
👋 @bmcfee & @heraldoborges - thanks for agreeing to review this! Your job now is to go through the review criteria in your checklist above, as discussed in the first 2 comments in this thread. If you can check things off, please do. When you see things that need to be addressed, if they are short, add comments here, and if they are more complex, add issues in the source repo and put the URL of this review thread there as well so this thread gets tagged.
If you have any questions or problems, please let me know.
