### 2020-01-06 #1995 @bmcfee
@danielskatz meta-question for you.  I've looked through the code of this package, and it appears that it works by web-scraping, and not through an officially sanctioned API.  IANAL, but I worry about whether the use of this package potentially violates terms of service, and whether it's appropriate for publication in JOSS given the strong dependency on (the benevolence of) a commercial entity.

To be clear, I'm not advocating one way or the other at this point, but AFAIK there is nothing in the JOSS reviewer guidelines that covers this kind of issue.  It would be nice to have some editorial clarification first before diving into the rest of the review.
