### 2020-01-07 #1995 @danielskatz
I don't see a problem with this, if there is no public/published API. If you think there is a better way to approach this, please suggest it.
