### 2020-01-07 #1995 @bmcfee
Genius does have an API, but it very intentionally does not provide lyrics data (which I'm sure they do not own), as far as I can tell.
