### 2020-01-07 #1995 @danielskatz
Ok, thanks - as long as you don't think there's a problem here in this software violating the genius terms of service, please continue the review
