### 2020-01-07 #1995 @bmcfee
> as long as you don't think there's a problem here in this software violating the genius terms of service,

Sorry, this is unclear to me.  Are you saying "as long as it's okay to violate ToS, proceed" or "as long as this software does not violate ToS, proceed"?

Going off recent news, I think I can confidently predict that this does violate ToS: https://www.theverge.com/2019/12/3/20993621/genius-google-lawsuit-stolen-lyrics-lyricfind

The question I have is what JOSS's stance on this kind of thing is, given that the project under review is very closely dependent on a proprietary service.  It would be another thing entirely for a general purpose scraper, eg beautifulsoup, but this specific project is very much about crawling content from one specific service, who would probably not be happy about it.
