### 2020-01-07 #1995 @danielskatz
I meant "as long as this software does not violate ToS, proceed"

But given what you have said, let's pause for a little bit and I'll discuss this with the other editors.
