### 2020-01-07 #1995 @jedbrown
The lawsuit appears to be primarily concerned with commercial use (see page 36 of [the filing](https://iapps.courts.state.ny.us/nyscef/ViewDocument?docIndex=3E0o8kQz4X3cWcbbid67wQ==)), but the terms related to scraping look relevant here (emphasis mine):

> Except as expressly authorized by Genius in writing, you agree not to modify, copy, frame, scrape, rent, lease, loan, sell, distribute or create derivative works based on the Service or the Genius Content, in whole or in part, except that the foregoing does not apply to your own User Content (as defined above) that you legally upload to the Service. In connection with your use of the Service you **shall not engage in or use any data mining, robots, scraping or similar data gathering or extraction methods**. -- https://genius.com/static/terms
