### 2020-01-08 #1995 @danielskatz
👋 @JosiahParry - The editors have decided to reject this submission due to the fact that the software appears to violate the terms of service for the service it uses (Genius).

@bmcfee & @heraldoborges - sorry for the unneeded work you may have done on this already, and @bmcfee, thanks specifically for noticing and bringing up this issue
