### 2020-01-08 #1995 @bmcfee
Thanks for the clarification @danielskatz , and sorry @josiahparry that this didn't work out!

@danielskatz - is there any discussion of getting some general verbiage in place to clarify these points (namely, tos violation and dependency on proprietary services) in the JOSS guidelines?  I'd be happy to help out with that if you think it would be useful.
