### 2020-01-08 #1995 @danielskatz
It would be great if you could suggest something in our docs via a PR, or an issue, in the [JOSS repo](https://github.com/openjournals/joss)
